+++ site/routes_tests.go
@@ -36,6 +36,7 @@ func setupTests(ctx context.Context, router chi.Router, signals sessions.Store) 
 				{ID: "local_signals"},
 				{ID: "merge_fragment"},
 				{ID: "merge_fragment_signal"},
+				{ID: "merge_fragment_whitespace"},
 				{ID: "on_load"},
 				{ID: "radio_input"},
 				{ID: "ref"},
@@ -97,6 +98,7 @@ func setupTests(ctx context.Context, router chi.Router, signals sessions.Store) 
 
 		if err := errors.Join(
 			setupTestsMergeFragment(testsRouter),
+			setupTestsMergeFragmentWhitespace(testsRouter),
 			setupTestsMergeFragmentSignal(testsRouter),
 			setupTestsOnLoad(testsRouter),
 			setupTestsRemoveFragment(testsRouter),