+++ site/smoketests/select_single_test.go
@@ -1,0 +1,9 @@
+package smoketests
+
+import (
+	"testing"
+)
+
+func TestSelectSingle(t *testing.T) {
+	setupPageTestOnSelect(t, "tests/select_single")
+}