+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Each tagged version of Datastar is accompanied by a release note. Read the [rele
 
 ## v1.0.0-beta.7
 
+### Changed
+
+- Updated Idiomorph to version [0.7.1](https://github.com/bigskysoftware/idiomorph/blob/main/CHANGELOG.md#071---2025-02-13).
+
 ### Fixed
 
 - Fixed a bug in which `datastar-remove-fragments` events were not having any effect ([#664](https://github.com/starfederation/datastar/issues/664)).