+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Each tagged version of Datastar is accompanied by a release note. Read the [rele
 ### Changed
 
 - Updated Idiomorph to version [0.7.1](https://github.com/bigskysoftware/idiomorph/blob/main/CHANGELOG.md#071---2025-02-13).
+- Whitespace is now maintained in merged fragments ([#658](https://github.com/starfederation/datastar/issues/658)).
 
 ### Fixed