+++ CHANGELOG.md
@@ -4,7 +4,11 @@ Each tagged version of Datastar is accompanied by a release note. Read the [rele
 
 # WIP Release Notes
 
-## v1.0.0-beta.7
+## v1.0.0-beta.8
+
+### Added
+
+- Added the ability for checkbox input elements to set bound signals to a boolean, string or array, depending on the predefined signal type ([#664](https://github.com/starfederation/datastar/issues/674)).
 
 ### Changed